api_server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
supabase==2.13.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic[email]==2.10.6
